dfpropパッケージのリファクタを行った
