+++ src/main/java/org/dbflute/intro/app/web/dfprop/syncschema/DfpropEditSyncSchemaBody.java
@@ -13,16 +13,22 @@
  * either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-package org.dbflute.intro.app.web.dfprop;
+package org.dbflute.intro.app.web.dfprop.syncschema;
 
-import javax.validation.constraints.NotNull;
+import org.lastaflute.web.validation.ClientError;
+import org.lastaflute.web.validation.Required;
 
 /**
- * @author hakiba
+ * @author deco
  */
-public class DfpropDeleteSchemaPolicyStatementBody {
-    @NotNull
-    public String mapType;
-    @NotNull
-    public String statement;
+public class DfpropEditSyncSchemaBody {
+
+    @Required
+    public String url;
+    public String schema;
+    @Required
+    public String user;
+    public String password;
+    @Required(groups = ClientError.class)
+    public Boolean isSuppressCraftDiff;
 }
