+++ src/static/app/common/factory/ApiFactory.js
@@ -3,13 +3,13 @@ export default class ApiFactory {
   //                                                                           Intro
   //                                                                           =====
   manifest() {
-    return ffetch.post('api/intro/manifest');
+    return ffetch.post("api/intro/manifest");
   }
   classifications() {
-    return ffetch.post('api/intro/classifications');
+    return ffetch.post("api/intro/classifications");
   }
   configuration() {
-    return ffetch.post('api/intro/configuration');
+    return ffetch.post("api/intro/configuration");
   }
 
   // ===============================================================================
@@ -17,119 +17,122 @@ export default class ApiFactory {
   //                                                                         =======
 
   createWelcomeClient(client, testConnection) {
-    return ffetch.post('api/welcome/create',
-      { body: { client: client, testConnection: testConnection } });
+    return ffetch.post("api/welcome/create", {
+      body: { client: client, testConnection: testConnection },
+    });
   }
 
   // ===============================================================================
   //                                                                          Client
   //                                                                          ======
   clientList() {
-    return ffetch.post('api/client/list');
+    return ffetch.post("api/client/list");
   }
   clientOperation(projectName) {
     return ffetch.post(`api/client/operation/${projectName}`);
   }
   createClient(client, testConnection) {
-    return ffetch.post(
-      `api/client/create`,
-      { body: { client, testConnection } }
-    )
+    return ffetch.post(`api/client/create`, {
+      body: { client, testConnection },
+    });
   }
   updateClient(clientBody, testConnection) {
-    return ffetch.post(`api/client/edit/${clientBody.projectName}`,
-      { body: { clientBody: clientBody, testConnection: testConnection } });
+    return ffetch.post(`api/client/edit/${clientBody.projectName}`, {
+      body: { clientBody: clientBody, testConnection: testConnection },
+    });
   }
   removeClient(clientBody) {
     return ffetch.post(`api/client/delete/${clientBody.project}`);
   }
   settings(projectName) {
-    return ffetch.post(`api/settings/${projectName}`)
+    return ffetch.post(`api/settings/${projectName}`);
   }
   updateSettings(clientBody) {
-    return ffetch.post(`api/settings/edit/${clientBody.projectName}`,
-      { body: { client: clientBody } });
+    return ffetch.post(`api/settings/edit/${clientBody.projectName}`, {
+      body: { client: clientBody },
+    });
   }
   dfporpBeanList(clientBody) {
-    return ffetch.post(`api/dfprop/${clientBody.projectName}/list`);
+    return ffetch.post(`api/dfprop/list/${clientBody.projectName}`);
   }
   syncSchema(projectName) {
-    return ffetch.post(`api/dfprop/${projectName}/syncschema`);
+    return ffetch.post(`api/dfprop/syncschema/${projectName}`);
   }
   editSyncSchema(projectName, syncSchemaSettingData) {
-    return ffetch.post(`api/dfprop/${projectName}/syncschema/edit`,
-      {
-        body: {
-          url: syncSchemaSettingData.url,
-          schema: syncSchemaSettingData.schema,
-          user: syncSchemaSettingData.user,
-          password: syncSchemaSettingData.password,
-          isSuppressCraftDiff: syncSchemaSettingData.isSuppressCraftDiff || false // need not null
-        }
-      });
+    return ffetch.post(`api/dfprop/syncschema/edit/${projectName}/`, {
+      body: {
+        url: syncSchemaSettingData.url,
+        schema: syncSchemaSettingData.schema,
+        user: syncSchemaSettingData.user,
+        password: syncSchemaSettingData.password,
+        isSuppressCraftDiff: syncSchemaSettingData.isSuppressCraftDiff || false, // need not null
+      },
+    });
   }
   schemaPolicy(projectName) {
-    return ffetch.post(`api/dfprop/${projectName}/schemapolicy`);
+    return ffetch.post(`api/dfprop/schemapolicy/${projectName}`);
   }
   editSchemaPolicy(projectName, schemaPolicyData) {
-    return ffetch.post(`api/dfprop/${projectName}/schemapolicy/edit`,
-      {
-        body: {
-          wholeMap: schemaPolicyData.wholeMap,
-          tableMap: schemaPolicyData.tableMap,
-          columnMap: schemaPolicyData.columnMap
-        }
-      });
+    return ffetch.post(`api/dfprop/schemapolicy/edit/${projectName}`, {
+      body: {
+        wholeMap: schemaPolicyData.wholeMap,
+        tableMap: schemaPolicyData.tableMap,
+        columnMap: schemaPolicyData.columnMap,
+      },
+    });
   }
   registerSchemapolicyStatement(projectName, schemaPolicyData) {
-    return ffetch.post(`api/dfprop/${projectName}/schemapolicy/statement/register`,
+    return ffetch.post(
+      `api/dfprop/schemapolicy/statement/register/${projectName}`,
       {
-        body: schemaPolicyData
-      });
+        body: schemaPolicyData,
+      }
+    );
   }
   getSchemapolicyStatementSubject() {
-    return ffetch.post(`api/dfprop/schemapolicy/statement/subject`)
+    return ffetch.post(`api/dfprop/schemapolicy/statement/subject`);
   }
   deleteSchemapolicyStatement(projectName, schemaPolicyData) {
-    return ffetch.post(`api/dfprop/${projectName}/schemapolicy/statement/delete`,
+    return ffetch.post(
+      `api/dfprop/schemapolicy/statement/delete/${projectName}`,
       {
-        body: schemaPolicyData
-      });
+        body: schemaPolicyData,
+      }
+    );
   }
   document(projectName) {
-    return ffetch.post(`api/dfprop/${projectName}/document`);
+    return ffetch.post(`api/dfprop/document/${projectName}`);
   }
   editDocument(projectName, documentSetting) {
-    return ffetch.post(`api/dfprop/${projectName}/document/edit`,
-      {
-        body: {
-          upperCaseBasic: documentSetting.upperCaseBasic,
-          aliasDelimiterInDbComment: documentSetting.aliasDelimiterInDbComment,
-          dbCommentOnAliasBasis: documentSetting.dbCommentOnAliasBasis,
-          checkColumnDefOrderDiff: documentSetting.checkColumnDefOrderDiff,
-          checkDbCommentDiff: documentSetting.checkDbCommentDiff,
-          checkProcedureDiff: documentSetting.checkProcedureDiff
-        }
-      });
+    return ffetch.post(`api/dfprop/document/edit/${projectName}`, {
+      body: {
+        upperCaseBasic: documentSetting.upperCaseBasic,
+        aliasDelimiterInDbComment: documentSetting.aliasDelimiterInDbComment,
+        dbCommentOnAliasBasis: documentSetting.dbCommentOnAliasBasis,
+        checkColumnDefOrderDiff: documentSetting.checkColumnDefOrderDiff,
+        checkDbCommentDiff: documentSetting.checkDbCommentDiff,
+        checkProcedureDiff: documentSetting.checkProcedureDiff,
+      },
+    });
   }
   openAlterDir(projectName) {
-    return ffetch.get(`api/playsql/migration/alter/${projectName}/open/`)
+    return ffetch.get(`api/playsql/migration/alter/open/${projectName}`);
   }
   alter(projectName) {
     return ffetch.get(`api/alter/${projectName}/`);
   }
   prepareAlterSql(projectName) {
-    return ffetch.post(`api/alter/${projectName}/prepare/`)
+    return ffetch.post(`api/alter/prepare/${projectName}/`);
   }
   createAlterSql(projectName, alterFileName) {
-    return ffetch.post(`api/alter/${projectName}/create/`, {
+    return ffetch.post(`api/alter/create/${projectName}/`, {
       body: {
-        alterFileName
-      }
-    })
+        alterFileName,
+      },
+    });
   }
   openDataDir(projectName) {
-    return ffetch.get(`api/playsql/data/${projectName}/open`)
+    return ffetch.get(`api/playsql/data/${projectName}/open`);
   }
   playsqlBeanList(projectName) {
     return ffetch.post(`api/playsql/${projectName}/list`);
@@ -142,23 +145,23 @@ export default class ApiFactory {
     return ffetch.post(`api/log`, {
       body: {
         project: projectName,
-        fileName: fileName
-      }
+        fileName: fileName,
+      },
     });
   }
 
   latestResult(projectName, task) {
-    return ffetch.get(`api/log/${projectName}/${task}/latest`)
+    return ffetch.get(`api/log/${projectName}/${task}/latest`);
   }
 
   // ===============================================================================
   //                                                                          Engine
   //                                                                          ======
   engineLatest() {
-    return ffetch.post('api/engine/latest');
+    return ffetch.post("api/engine/latest");
   }
   engineVersions() {
-    return ffetch.post('api/engine/versions');
+    return ffetch.post("api/engine/versions");
   }
   // needs trailing slash if URL parameter contains dot
   downloadEngine(params) {
@@ -179,11 +182,11 @@ export default class ApiFactory {
   //                                                                           =====
   retry(method, url, data, useSystemProxies) {
     data = data || {};
-    data['useSystemProxies'] = useSystemProxies;
+    data["useSystemProxies"] = useSystemProxies;
     return http({
       method: method,
       url: url,
-      data: data
+      data: data,
     });
   }
 }
